third_party/blink/web_tests/platform/win7/virtual/threaded/external/wpt/css/css-scroll-snap/snap-after-relayout/resnap-to-focused-expected.txt

This is a testharness.js-based test.
FAIL Resnap to focused element after relayout assert_equals: After resize, should snap to row 3. expected 3 but got 4
Harness: the test ran to completion.
